Baseline/requirements.txt
openai>=1.12.0
dataclasses-json>=0.6.4
pytest>=8.0.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
